+++ src/hooks/useOrder.ts
@@ -1,11 +1,8 @@
-import { useEffect, useState } from "react";
-import { fetchAddOrderList, fetchOrderList } from "../api";
-import { LocalProductType, OrderItemList } from "../types/domain";
+import { fetchAddOrderList } from "../api";
+import { LocalProductType } from "../types/domain";
 
 export const useOrder = () => {
-  const [orderList, setOrderList] = useState<OrderItemList[]>([]);
-
-  const addOrderList = async (
+  const addOrder = async (
     orderList: LocalProductType[],
     couponId: number | null
   ) => {
@@ -18,23 +15,5 @@ export const useOrder = () => {
     }
   };
 
-  const getOrderList = async () => {
-    try {
-      const response = await fetchOrderList();
-
-      if (!response.ok) throw new Error(response.status.toString());
-
-      const orderListData = await response.json();
-
-      setOrderList(orderListData.orders);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  useEffect(() => {
-    getOrderList();
-  }, []);
-
-  return { addOrderList, orderList };
+  return { addOrder };
 };
